chess_game/requirements.txt
python-chess==1.999
Pillow==10.1.0
requests==2.31.0
